Move e.i.l.c.util.StringUtils.trimXMP() to e.i.l.c.image.xmp.Utils
