+++ src/main/java/edu/illinois/library/cantaloupe/image/xmp/Utils.java
@@ -28,4 +28,17 @@ public final class Utils {
         return b.toString();
     }
 
+    /**
+     * Strips any enclosing tags or other content around the {@literal rdf:RDF}
+     * element within an RDF/XML XMP string.
+     */
+    public static String trimXMP(String xmp) {
+        final int start = xmp.indexOf("<rdf:RDF");
+        final int end = xmp.indexOf("</rdf:RDF");
+        if (start > -1 && end > -1) {
+            xmp = xmp.substring(start, end + 10);
+        }
+        return xmp;
+    }
+
 }
