+++ src/main/java/edu/illinois/library/cantaloupe/processor/codec/jpeg2000/JPEG2000KakaduImageReader.java
@@ -2,11 +2,11 @@ package edu.illinois.library.cantaloupe.processor.codec.jpeg2000;
 
 import edu.illinois.library.cantaloupe.image.Rectangle;
 import edu.illinois.library.cantaloupe.image.ScaleConstraint;
+import edu.illinois.library.cantaloupe.image.xmp.Utils;
 import edu.illinois.library.cantaloupe.operation.ReductionFactor;
 import edu.illinois.library.cantaloupe.operation.Scale;
 import edu.illinois.library.cantaloupe.processor.SourceFormatException;
 import edu.illinois.library.cantaloupe.util.Stopwatch;
-import edu.illinois.library.cantaloupe.util.StringUtils;
 import kdu_jni.Jp2_threadsafe_family_src;
 import kdu_jni.Jpx_codestream_source;
 import kdu_jni.Jpx_input_box;
@@ -527,7 +527,7 @@ public final class JPEG2000KakaduImageReader implements AutoCloseable {
                                 iptc = Arrays.copyOfRange(buffer, 16, buffer.length);
                             } else if (isXMP) {
                                 xmp = new String(buffer, StandardCharsets.UTF_8);
-                                xmp = StringUtils.trimXMP(xmp);
+                                xmp = Utils.trimXMP(xmp);
                             }
                         } finally {
                             box.Close();
