+++ src/main/java/edu/illinois/library/cantaloupe/util/StringUtils.java
@@ -218,19 +218,6 @@ public final class StringUtils {
     }
 
     /**
-     * Strips any enclosing tags or other content around the {@literal rdf:RDF}
-     * element within an RDF/XML XMP string.
-     */
-    public static String trimXMP(String xmp) {
-        final int start = xmp.indexOf("<rdf:RDF");
-        final int end = xmp.indexOf("</rdf:RDF");
-        if (start > -1 && end > -1) {
-            xmp = xmp.substring(start, end + 10);
-        }
-        return xmp;
-    }
-
-    /**
      * Returns an array of strings, one for each line in the string after it
      * has been wrapped to fit lines of <var>maxWidth</var>. Lines end with any
      * of CR, LF, or CRLF. A line ending at the end of the string will not
