+++ src/test/java/edu/illinois/library/cantaloupe/image/xmp/UtilsTest.java
@@ -17,4 +17,32 @@ class UtilsTest extends BaseTest {
         assertTrue(actual.endsWith("<?xpacket end=\"r\"?>"));
     }
 
+    /* trimXMP() */
+
+    @Test
+    void testTrimXMPWithTrimmableXMP() {
+        String xmp = "<?xpacket id=\"cats\"?>" +
+                "<x:xmpmeta bla=\"dogs\">" +
+                "<rdf:RDF foxes=\"bugs\">" +
+                "</rdf:RDF>" +
+                "</x:xmpmeta>";
+        String result = Utils.trimXMP(xmp);
+        assertTrue(result.startsWith("<rdf:RDF"));
+        assertTrue(result.endsWith("</rdf:RDF>"));
+    }
+
+    @Test
+    void testTrimXMPWithNonTrimmableXMP() {
+        String xmp = "<rdf:RDF foxes=\"bugs\">" +
+                "</rdf:RDF>";
+        String result = Utils.trimXMP(xmp);
+        assertSame(xmp, result);
+    }
+
+    @Test
+    void testTrimXMPWithNullArgument() {
+        assertThrows(NullPointerException.class,
+                () -> Utils.trimXMP(null));
+    }
+
 }
